# docs/swagger.yaml
basePath: /
definitions:
...
  consts.CardType:
    enum:
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - StudentCard
    - NormalCard
    - VipCard
  consts.Role:
    enum:
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - AdminRole
    - UserRole
    - StaffRole
  consts.Status:
    enum:
    - active
    - inactive
    - blocked
    type: string
    x-enum-varnames:
    - ActiveStatus
    - InactiveStatus
    - BlockedStatus
...
  models.Card:
    properties:
      balance:
        type: number
      created_at:
        type: string
      id:
        type: integer
      price:
        type: number
      rf_id:
        type: string
      status:
        $ref: '#/definitions/consts.Status'
      type:
        $ref: '#/definitions/consts.CardType'
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: quan hệ belongs-to
      username:
        description: 1-1 nên unique
        type: string
    type: object
...
  models.SellHistory:
    properties:
      card:
        allOf:
        - $ref: '#/definitions/models.Card'
        description: Foreign key relationships
      card_id:
        type: string
      card_price_sold:
        type: number
      created_at:
        type: string
      id:
        type: integer
      seller:
        $ref: '#/definitions/models.User'
      seller_id:
        type: string
      time:
        type: string
      updated_at:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      role:
        allOf:
        - $ref: '#/definitions/consts.Role'
        description: '"ADMIN", "USER", "STAFF"'
      status:
        description: '"active" hoặc "inactive"'
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A Metro Card Management System API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Metro API
  version: "1.0"
paths:
...
  /sell-history:
    get:
      consumes:
      - application/json
      description: Retrieve all sell history records with optional filtering
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of records per page
        in: query
        name: limit
        type: integer
      - description: Filter by card ID
        in: query
        name: card_id
        type: string
      - description: Filter by seller ID
        in: query
        name: seller_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sell histories retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.SellHistory'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get all sell history records
      tags:
      - sell-history
  /sell-history/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific sell history record by its ID
      parameters:
      - description: Sell History ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sell history retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SellHistory'
              type: object
        "404":
          description: Sell history not found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get sell history by ID
      tags:
      - sell-history
  /sell-history/card/{card_id}:
    get:
      consumes:
      - application/json
      description: Retrieve all sell history records for a specific card
      parameters:
      - description: Card ID
        in: path
        name: card_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sell histories retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.SellHistory'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get sell histories by card ID
      tags:
      - sell-history
  /sell-history/seller/{seller_id}:
    get:
      consumes:
      - application/json
      description: Retrieve all sell history records for a specific seller
      parameters:
      - description: Seller ID
        in: path
        name: seller_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sell histories retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.SellHistory'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get sell histories by seller ID
      tags:
      - sell-history
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"